+++ src/neuroconv/datainterfaces/ecephys/spike2/spike2datainterface.py
@@ -19,9 +19,7 @@ class Spike2RecordingInterface(BaseRecordingExtractorInterface):
     Data interface class for converting Spike2 data from CED (Cambridge Electronic
     Design) using the :py:class:`~spikeinterface.extractors.CedRecordingExtractor`."""
 
-    keywords = BaseRecordingExtractorInterface.keywords + [
-        "CED",
-    ]
+    keywords = BaseRecordingExtractorInterface.keywords + ["CED"]
 
     ExtractorName = "CedRecordingExtractor"
 
@@ -59,22 +57,3 @@ class Spike2RecordingInterface(BaseRecordingExtractorInterface):
         signal_channels = self.recording_extractor.neo_reader.header["signal_channels"]
         channel_ids_of_raw_data = [channel_info[1] for channel_info in signal_channels if channel_info[4] == "mV"]
         self.recording_extractor = self.recording_extractor.channel_slice(channel_ids=channel_ids_of_raw_data)
-
-
-class CEDRecordingInterface(Spike2RecordingInterface):
-    def __init__(self, file_path: FilePathType, verbose: bool = True, es_key: str = "ElectricalSeries"):
-        """
-        Initialize reading of CED file.
-
-        Parameters
-        ----------
-        file_path : FilePathType
-            Path to .smr or .smrx file.
-        verbose : bool, default: True
-        es_key : str, default: "ElectricalSeries"
-        """
-        warn(
-            message="CEDRecordingInterface will soon be deprecated. Please use Spike2RecordingInterface instead.",
-            category=DeprecationWarning,
-        )
-        super().__init__(file_path=file_path, verbose=verbose, es_key=es_key)