+++ src/neuroconv/tools/nwb_helpers/_metadata_and_file_helpers.py
@@ -1,4 +1,4 @@
-import json
+"""Collection of helper functions related to NWB."""
 import uuid
 from contextlib import contextmanager
 from copy import deepcopy
@@ -7,13 +7,12 @@ from pathlib import Path
 from typing import Optional
 from warnings import warn
 
-import jsonschema
+from pydantic import FilePath
 from pynwb import NWBHDF5IO, NWBFile
 from pynwb.file import Subject
 
-from ..utils import FilePathType, dict_deep_update
-from ..utils.dict import DeepDict, load_dict_from_file
-from ..utils.json_schema import validate_metadata
+from ...utils.dict import DeepDict, load_dict_from_file
+from ...utils.json_schema import validate_metadata
 
 
 def get_module(nwbfile: NWBFile, name: str, description: str = None):
@@ -51,9 +50,8 @@ def get_default_nwbfile_metadata() -> DeepDict:
 
 def make_nwbfile_from_metadata(metadata: dict) -> NWBFile:
     """Make NWBFile from available metadata."""
-
     # Validate metadata
-    schema_path = Path(__file__).resolve().parent.parent / "schemas/base_metadata_schema.json"
+    schema_path = Path(__file__).resolve().parent.parent.parent / "schemas" / "base_metadata_schema.json"
     base_metadata_schema = load_dict_from_file(file_path=schema_path)
     validate_metadata(metadata=metadata, schema=base_metadata_schema)
 
@@ -130,7 +128,7 @@ def add_device_from_metadata(nwbfile: NWBFile, modality: str = "Ecephys", metada
 
 @contextmanager
 def make_or_load_nwbfile(
-    nwbfile_path: Optional[FilePathType] = None,
+    nwbfile_path: Optional[FilePath] = None,
     nwbfile: Optional[NWBFile] = None,
     metadata: Optional[dict] = None,
     overwrite: bool = False,